p15267807 Products admin display loves and comments   + Order by most loved   + Order by most commented

+++ features/step_definitions/product_steps.rb
@@ -24,10 +24,31 @@ Given /^I (only )?have products titled "?([^\"]*)"?$/ do |only, titles|
 end
 
 Given /^I only have a foo product$/ do
-  Product.delete_all 
+  Product.delete_all
   Given %{I have a product titled "foo product"}
+end
+
+Given /^I only have a foo product with ([^\d]*)$/ do
+  Product.delete_all
+  Given %{I have a product titled "foo product"}
+end
+
+Then /^I (only )?have a ([^\"]*) product with (\d+) love[s]? and (\d+) comment[s]?$/ do |only, name, loves, comments|
+  Product.delete_all if only
+  Given %{I have a product titled "#{name} product"}
+  product = Product.where(:name => "#{name} product").first
+  loves.to_i.times do
+    product.love.create
+  end
+  comments.to_i.times do
+    product.comments.create!(:message => "test comment")
+  end
 end
 
 Then /^I should have ([0-9]+) products?$/ do |count|
   Product.count.should == count.to_i
+end
+
+Then /^"([^"]*)" product should appear before "([^"]*)"$/ do |first_product,second_product|
+  page.body.should =~ /.*#{first_product}.*#{second_product}/m
 end
